Enable ci_job_trace_force_encode feature flag by default
In the oft chance that canceled builds need to append to object storage
for the CI job trace, this feature flag ensures the data can be appended
properly with the right encoding and that jobs will actually end
properly.

We saw this fixed an issue on GitLab.com immediately after it was
enabled.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/333452

Changelog: changed