Check if default branch is already protected while creating
This change checks if the default branch is already
protected before creating a protected branch. This happens when a
security policy project is created with branch protection enabled
for the default branch. To avoid the race condition from
BranchHooksService we update the branch protection if it
already exists while creating security policy project.