Bugfix: Disable the ability from the FE to use Approver != in MR search
This functionality currently does not work, and the query will be very
expensive and complex to do. So in the meantime we should remove the
ability to search for it from the frontend to not mislead customers.

Issue: gitlab.com/gitlab-org/gitlab/-/issues/218570
MR: gitlab.com/gitlab-org/gitlab/-/merge_requests/56310

Changelog: fixed