Ensure updating browsers forces babel to recompile
Given that we use `@babel/preset-env` transpiled files contain imports
to `core-js` at the top of the file. `babel-loader` caches those
transpiled files. If we bump the supported browser versions in
`.browserslistrc`, we might need less imports of `corejs` modules. But
as they are cached, we will retrieve those outdated files with the wrong
imports causing unneccessary `core-js` modules to be imported.