Fix LFS objects not downloading from Bitbucket
Previously attempting to download LFS files with Bitbucket via a project
import would result in a 401 error. This happened because Bitbucket
returns a `header` payload that should be included in the HTTP GET
request. As described in
https://github.com/git-lfs/git-lfs/blob/main/docs/api/batch.md, we need
to include this with the request for the download to succeed.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/335123

Changelog: fixed