Automatically create integration webhooks when missing
Previously we were silently failing if the webhook could not be saved
when the integration was edited, which led to further exceptions later
when the integration was trying to execute the webhook.

The webhook is really only used to associate records in `web_hook_logs`,
so we can create it automatically when needed and don't have to
backfill existing integrations.

We now also raise errors if the webhook cannot be saved, which should
help us diagnose why this is failing in the first place.

Changelog: fixed