Correct check for displaying SubEpic counts on boards
The `allowSubEpics` state variable was not being set previously so any
checks in app/assets/javascripts/boards/components/board_card_inner.vue
would be checking against `undefined`. We need to set the data on the
board app and add it to the state as the page loads.

Changelog: fixed
EE: true