Allow configuring Redis trace chunks instance
Allow configuring a Redis instance for storing trace chunks. For almost
every GitLab instance except GitLab.com, all Redis instances are the
same instance. The trace chunks REdis is only used in the
`ci_enable_live_trace` feature flag is enabled (the flag is off by
default).

If that flag is enabled and the instance was already using the
persistent Redis, then the only way to cleanly migrate is to take
downtime. We are not providing additional migration support at this time
as this is a niche use-case on top of a niche use-case, and so the
additional complexity in the application isn't worth it right now.

Changelog: added