Fix filePath not being stored in GraphQL log tree cache
This was discovered when trying to paginate the tree entries and use the
log tree cache: we were not storing `filePath` even though it was being
generated. As a result, the log entries could not be matched with
newly-inserted data.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/334140

Changelog: fixed