Drive container registry migration phase 1 with Rails FFs
Adds a series of FFs to control and gradually increase the number of
new container repositories that will use the upcoming container
registry metadata database for GitLab.com.

These feature flags will then be used to modify the JWT tokens served
by Rails to clients whenever they want to interact with the registry.
The registry will then read the information within those tokens and
decide if a new repository should use the database or not.